requirements.txt
certifi>=2020.11.8
chardet>=3.0.4
idna>=2.10
Jinja2>=2.11.2
Mako>=1.1.3
MarkupSafe>=1.1.1
PyYAML>=5.3.1
requests>=2.25.0
tabulate>=0.8.7
urllib3>=1.26.2
